LayoutTests/imported/w3c/web-platform-tests/css/css-logical/logical-values-resize-expected.txt


FAIL Test that 'resize: block' is supported. assert_equals: logical values in inline style, resize expected "block" but got ""
FAIL Test that 'resize: inline' is supported. assert_equals: logical values in inline style, resize expected "inline" but got ""
